+++ frontend/src/routes/ops/+page.svelte
@@ -2,18 +2,15 @@
   import { Tabs, TabItem, Button, Toast } from 'flowbite-svelte';
   import { ArchiveSolid, ArrowDownToBracketOutline, CloseCircleSolid, CheckCircleSolid, ClipboardListSolid, TableRowSolid,PlusOutline, ChevronDownOutline, FilterSolid, ChevronRightOutline, ChevronLeftOutline } from 'flowbite-svelte-icons';
   import { Section } from 'flowbite-svelte-blocks';
-  import { Table, TableBody, TableBodyCell, TableBodyRow, TableHead, TableHeadCell, TableSearch, Dropdown, DropdownItem, Checkbox, ButtonGroup } from 'flowbite-svelte';
+  import { Table, TableBody, TableBodyCell, TableBodyRow, TableHead, TableHeadCell } from 'flowbite-svelte';
 	import { onMount } from 'svelte';
-
+  import { DataTable } from '$lib/components/DataTable.svelte';
   import { Timeline, TimelineItem, Activity, ActivityItem, Group, GroupItem } from 'flowbite-svelte';
   import { invoke } from '@tauri-apps/api/tauri';
   import { fly } from 'svelte/transition';
   
   import { Input, Label } from 'flowbite-svelte';
-let paginationData;
-  onMount(async () =>{
-    paginationData= await invoke('recall_all_hashes');
-  });
+
   let result: string | null = null;
   let counter = 5;
   let toastStatus = false;
@@ -62,23 +59,7 @@ let paginationData;
   }
   let json = '';
   let tags = {};
-
-  let divClass='bg-white dark:bg-gray-800 relative shadow-md sm:rounded-lg overflow-hidden';
-  let innerDivClass='flex flex-col md:flex-row items-center justify-between space-y-3 md:space-y-0 md:space-x-4 p-4';
-  let searchClass='w-full md:w-1/2 relative';
-  let svgDivClass='absolute inset-y-0 left-0 flex items-center pl-3 pointer-events-none';
-  let classInput="text-gray-900 text-sm rounded-lg focus:ring-primary-500 focus:border-primary-500 block w-full p-2  pl-10";
-
-  let searchTerm = '';
-  let currentPosition = 0;
-  const itemsPerPage = 10;
-  const showPage = 5;
-  let totalPages = 0;
-  let pagesToShow = [];
-  let totalItems;
-  let startPage;
-  let endPage;
-  
+  let paginationData;
   async function getTop5TableData(){
     console.log('running getTableData function');
     invoke('log_message', {message:"Running get table data function"});
@@ -90,59 +71,11 @@ let paginationData;
   async function getTableData(){
     console.log('running getTableData function');
     invoke('log_message', {message:"Running get table data function"});
-    //paginationData = await invoke('recall_all_hashes');
+    paginationData = await invoke('recall_all_hashes');
     tags = await invoke('recall_tags');
-    totalItems = paginationData.length;
     // time to reformat the json so its
-    renderPagination(paginationData.length);
-
-  }
-  const updateDataAndPagination = () => {
-    const currentPageItems = paginationData.slice(currentPosition, currentPosition + itemsPerPage);
-    renderPagination(currentPageItems.length);
   }
 
-  const loadNextPage = () => {
-    if (currentPosition + itemsPerPage < paginationData.length) {
-      currentPosition += itemsPerPage;
-      updateDataAndPagination();
-    }
-  }
-
-  const loadPreviousPage = () => {
-    if (currentPosition - itemsPerPage >= 0) {
-      currentPosition -= itemsPerPage;
-      updateDataAndPagination();
-    }
-  }
-
-  const renderPagination = (totalItems) => {
-    totalPages = Math.ceil(paginationData.length / itemsPerPage);
-    const currentPage = Math.ceil((currentPosition + 1) / itemsPerPage);
-
-    startPage = currentPage - Math.floor(showPage / 2);
-    startPage = Math.max(1, startPage);
-    endPage = Math.min(startPage + showPage - 1, totalPages);
-
-    pagesToShow = Array.from({ length: endPage - startPage + 1 }, (_, i) => startPage + i);
-  }
-
-  const goToPage = (pageNumber) => {
-    currentPosition = (pageNumber - 1) * itemsPerPage;
-    updateDataAndPagination();
-  }
-  //getTableData();
-  $: startRange = currentPosition + 1;
-  $: endRange = Math.min(currentPosition + itemsPerPage, totalItems);
-
-  $: currentPageItems = paginationData.slice(currentPosition, currentPosition + itemsPerPage);
-  $: filteredItems = paginationData.filter((item) => {
-  const searchTermLower = searchTerm.toLowerCase();
-  return (
-    item.title.toLowerCase().indexOf(searchTermLower) !== -1 ||
-    item.hash.toLowerCase().indexOf(searchTermLower) !== -1
-  );
-});
 </script>
 
 <html class="dark" lang="en">
@@ -226,78 +159,17 @@ let paginationData;
     {/each}
     
   </TableBody>
+
 </Table>
-        <Section name="advancedTable" classSection='bg-gray-50 dark:bg-gray-900 p-3 sm:p-5' on:load={getTableData()}>
-    <TableSearch placeholder="Search" hoverable={true} bind:inputValue={searchTerm} {divClass} {innerDivClass} {searchClass} {classInput} >
-
-    <div slot="header" class="w-full md:w-auto flex flex-col md:flex-row space-y-2 md:space-y-0 items-stretch md:items-center justify-end md:space-x-3 flex-shrink-0">
-      <Button>
-        <PlusOutline class="h-3.5 w-3.5 mr-2" />Add item
-      </Button>
-      
-      <Button color='alternative'>Filter<FilterSolid class="w-3 h-3 ml-2 " /></Button>
-        <Dropdown class="w-48 p-3 space-y-2 text-sm">
-          <h6 class="mb-3 text-sm font-medium text-gray-900 dark:text-white">Choose category</h6>
-          {#each tags as tag}
-          <li>
-            <Checkbox>{tag.name}</Checkbox>
-          </li>
-          {/each}
-          
-        </Dropdown>
-    </div>
-      <TableHead>
-        <TableHeadCell padding="px-4 py-3" scope="col">ID</TableHeadCell>
-        <TableHeadCell padding="px-4 py-3" scope="col">Hash</TableHeadCell>
-        <TableHeadCell padding="px-4 py-3" scope="col">Title</TableHeadCell>
-        <TableHeadCell padding="px-4 py-3" scope="col">Description</TableHeadCell>
-        <TableHeadCell padding="px-4 py-3" scope="col">Tag</TableHeadCell>
-        <TableHeadCell padding="px-4 py-3" scope="col">Time</TableHeadCell>
-      </TableHead>
-      <TableBody class="divide-y">
-        {#if searchTerm !== ''}
-          {#each filteredItems as item (item.id)}
-            <TableBodyRow>
-              <TableBodyCell tdClass="px-4 py-3">{item.id}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.hash}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.title}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.description}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.tag}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.time_unix}</TableBodyCell>
-            </TableBodyRow>
-          {/each}
-        {:else}
-          {#each currentPageItems as item (item.id)}
-            <TableBodyRow>
-              <TableBodyCell tdClass="px-4 py-3">{item.id}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.hash}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.title}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.description}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.tag}</TableBodyCell>
-              <TableBodyCell tdClass="px-4 py-3">{item.time_unix}</TableBodyCell>
-            </TableBodyRow>
-          {/each}
-        {/if}
-      </TableBody>
-      <div slot="footer" class="flex flex-col md:flex-row justify-between items-start md:items-center space-y-3 md:space-y-0 p-4" aria-label="Table navigation">
-      <span class="text-sm font-normal text-gray-500 dark:text-gray-400">
-        Showing
-        <span class="font-semibold text-gray-900 dark:text-white">{startRange}-{endRange}</span>
-        of
-        <span class="font-semibold text-gray-900 dark:text-white">{totalItems}</span>
-      </span>
-        <ButtonGroup>
-          <Button on:click={loadPreviousPage} disabled={currentPosition === 0}><ChevronLeftOutline size='xs' class='m-1.5'/></Button>
-          {#each pagesToShow as pageNumber}
-            <Button on:click={() => goToPage(pageNumber)}>{pageNumber}</Button>
-          {/each}
-          <Button on:click={loadNextPage} disabled={ totalPages === endPage }><ChevronRightOutline size='xs' class='m-1.5'/></Button>
-        </ButtonGroup>
-      </div>
-    </TableSearch>
-</Section> 
-
-      </TabItem>
+  {#if paginationData != null && tags != {}}
+ <DataTable paginationData=paginationData tags=tags />
+  {:else}
+  <Label class="space-y-2">
+      <br>
+      <Input type="text" placeholder="MatrixFS Hash" size="md" disabled value="Loading Data..." />
+  </Label> 
+  {/if}
+</TabItem>
       <TabItem>
         <div slot="title" class="flex items-center gap-2">
           <ClipboardListSolid class="w-5 h-5"/>
